requirements.txt
anthropic
mss
Pillow
python-telegram-bot
schedule
python-dotenvyfinance
google-generativeai
google-genai
pytz
